+++ cashshop/handle_client.go
@@ -6,9 +6,11 @@ import (
 	"github.com/Hucaru/Valhalla/channel"
 	"github.com/Hucaru/Valhalla/common"
 	"github.com/Hucaru/Valhalla/common/opcode"
+	"github.com/Hucaru/Valhalla/constant"
 	"github.com/Hucaru/Valhalla/internal"
 	"github.com/Hucaru/Valhalla/mnet"
 	"github.com/Hucaru/Valhalla/mpacket"
+	"github.com/Hucaru/Valhalla/nx"
 )
 
 func (server *Server) HandleClientPacket(conn mnet.Client, reader mpacket.Reader) {
@@ -25,37 +27,6 @@ func (server *Server) HandleClientPacket(conn mnet.Client, reader mpacket.Reader
 
 	default:
 		log.Println("UNKNOWN CASHSHOP PACKET (", op, "):", reader)
-	}
-}
-
-func (server *Server) HandleServerPacket(conn mnet.Server, reader mpacket.Reader) {
-	switch reader.ReadByte() {
-	case opcode.ChannelPlayerConnect:
-
-	case opcode.ChannePlayerDisconnect:
-		server.handlePlayerDisconnect(conn, reader)
-	case opcode.ChannelConnectionInfo:
-		server.handleChannelConnectionInfo(conn, reader)
-	case opcode.CashShopOk:
-		server.handleWorldConnection(conn, reader)
-	case opcode.CashShopBad:
-		log.Panicln("CashShop unable to connect to world")
-	default:
-		log.Println("UNKNOWN SERVER PACKET:", reader)
-	}
-}
-
-func (server Server) handleWorldConnection(conn mnet.Server, reader mpacket.Reader) {
-	reader.ReadString(reader.ReadInt16())
-	log.Printf("Connected to world %s\n", conn)
-}
-
-func (server *Server) handleChannelConnectionInfo(conn mnet.Server, reader mpacket.Reader) {
-	total := reader.ReadByte()
-
-	for i := range total {
-		server.channels[i].IP = reader.ReadBytes(4)
-		server.channels[i].Port = reader.ReadInt16()
 	}
 }
 
@@ -110,28 +81,6 @@ func (server *Server) handlePlayerConnect(conn mnet.Client, reader mpacket.Reade
 	plr.Send(packetCashShopSet(&plr))
 	plr.Send(packetCashShopUpdateAmounts(plr.GetNX(), plr.GetMaplePoints()))
 	plr.Send(packetCashShopWishList(nil, true))
-}
-
-func (server *Server) handlePlayerDisconnect(conn mnet.Server, reader mpacket.Reader) {
-	playerID := reader.ReadInt32()
-	name := reader.ReadString(reader.ReadInt16())
-	_ = reader.ReadInt32()
-
-	plr, err := server.players.GetFromID(playerID)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-	err = server.players.RemoveFromConn(plr.Conn)
-	if err != nil {
-		return
-	}
-
-	if _, err := common.DB.Exec("UPDATE characters SET inCashShop=0 WHERE ID=?", playerID); err != nil {
-		return
-	}
-
-	log.Printf("Player %s (ID: %d) disconnected from CashShop\n", name, playerID)
 }
 
 func (server *Server) leaveCashShopToChannel(conn mnet.Client, reader mpacket.Reader) {
@@ -192,4 +141,124 @@ func (server *Server) leaveCashShopToChannel(conn mnet.Client, reader mpacket.Re
 	p.WriteBytes(ip)
 	p.WriteInt16(port)
 	conn.Send(p)
+}
+
+func (server *Server) playerCashShopPurchase(conn mnet.Client, reader mpacket.Reader) {
+	plr, err := server.players.GetFromConn(conn)
+	if err != nil {
+		return
+	}
+
+	plrNX := plr.GetNX()
+	plrMaplePoints := plr.GetMaplePoints()
+
+	sub := reader.ReadByte()
+	switch sub {
+	case opcode.RecvCashShopBuyItem:
+		currencySel := reader.ReadByte()
+		sn := reader.ReadInt32()
+
+		commodity, ok := nx.GetCommodity(sn)
+		if !ok || commodity.ItemID == 0 {
+			// Unknown SN
+			plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+			return
+		}
+
+		if commodity.OnSale == 0 || commodity.Price <= 0 {
+			plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+			return
+		}
+
+		// Determine quantity
+		count := int16(1)
+		if commodity.Count > 0 {
+			count = int16(commodity.Count)
+		}
+
+		// Check funds
+		price := commodity.Price
+		switch currencySel {
+		case constant.CashShopNX:
+			if plrNX < price {
+				plr.Send(packetCashShopError(opcode.SendCashShopBuyFailed, constant.CashShopErrorNotEnoughCash))
+				return
+			}
+		case constant.CashShopMaplePoints:
+			if plrMaplePoints < price {
+				plr.Send(packetCashShopError(opcode.SendCashShopBuyFailed, constant.CashShopErrorNotEnoughCash))
+				return
+			}
+		default:
+			plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+			return
+		}
+
+		newItem, e := channel.CreateItemFromID(commodity.ItemID, count)
+		if e != nil {
+			plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+			return
+		}
+
+		if err, _ := plr.GiveItem(newItem); err != nil {
+			plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+			return
+		}
+
+		switch currencySel {
+		case constant.CashShopNX:
+			plrNX -= price
+			plr.SetNX(plrNX)
+		case constant.CashShopMaplePoints:
+			plrMaplePoints -= price
+			plr.SetMaplePoints(plrMaplePoints)
+		default:
+			log.Println("Unknown currency type: ", currencySel)
+			return
+		}
+
+		plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+
+	case opcode.RecvCashShopGiftItem:
+	case opcode.RecvCashShopUpdateWishlist:
+	case opcode.RecvCashShopIncreaseSlots:
+		currencySel := reader.ReadByte()
+		invType := reader.ReadByte()
+
+		price := int32(4000)
+
+		switch currencySel {
+		case constant.CashShopNX:
+			if plrNX < price {
+				plr.Send(packetCashShopError(opcode.SendCashShopIncSlotCountFailed, constant.CashShopErrorNotEnoughCash))
+				return
+			}
+			if err := plr.IncreaseSlotSize(invType, 4); err != nil {
+				plr.Send(packetCashShopError(opcode.SendCashShopIncSlotCountFailed, constant.CashShopErrorUnknown))
+				return
+			}
+			plrNX -= price
+			plr.SetNX(plrNX)
+		case constant.CashShopMaplePoints:
+			if plrMaplePoints < price {
+				plr.Send(packetCashShopError(opcode.SendCashShopIncSlotCountFailed, constant.CashShopErrorNotEnoughCash))
+				return
+			}
+			if err := plr.IncreaseSlotSize(invType, 4); err != nil {
+				plr.Send(packetCashShopError(opcode.SendCashShopIncSlotCountFailed, constant.CashShopErrorUnknown))
+				return
+			}
+			plrMaplePoints -= price
+			plr.SetMaplePoints(plrMaplePoints)
+		default:
+			log.Println("Unknown currency type: ", currencySel)
+			return
+		}
+
+		plr.Send(packetCashShopIncreaseInv(invType, plr.GetSlotSize(invType)))
+		plr.Send(packetCashShopUpdateAmounts(plrNX, plrMaplePoints))
+	default:
+		log.Println("Unknown Cash Shop Packet(", sub, "): ", reader)
+	}
+
 }
